Moved socket io functions to another file
